Reorg code to different files
Build ok

Lint
